+++ api/src/main/java/com/coinbase/exchange/api/exchange/CoinbaseExchange.java
@@ -10,9 +10,9 @@ public interface CoinbaseExchange {
     String getBaseUrl();
     <R> HttpEntity<String> securityHeaders(String endpoint, String method, String body);
     <T> T get(String endpoint, ParameterizedTypeReference<T> type);
-    <T> T pagedGet(String endpoint, ParameterizedTypeReference<T> responseType, String beforeOrAfter, Integer pageNumber, Integer limit);
+    <T> T pagedGet(String endpoint, ParameterizedTypeReference<T> responseType, String beforeOrAfter, Long pageNumber, Integer limit);
     <T> List<T> getAsList(String endpoint, ParameterizedTypeReference<T[]> type);
-    <T> List<T> pagedGetAsList(String endpoint, ParameterizedTypeReference<T[]> responseType, String beforeOrAfter, Integer pageNumber, Integer limit);
+    <T> List<T> pagedGetAsList(String endpoint, ParameterizedTypeReference<T[]> responseType, String beforeOrAfter, Long pageNumber, Integer limit);
     <T, R> T post(String endpoint, ParameterizedTypeReference<T> type, R jsonObject);
     <T> T delete(String endpoint, ParameterizedTypeReference<T> type);
 }
